feat: swr for useStreamStatus & refactor more components
