+++ apps/app/components/daydream/MainExperience/index.tsx
@@ -1,10 +1,11 @@
-import DayDreamContent from "./DaydreamContent";
-import { useOnboard } from "../OnboardContext";
+import ClientSideTracker from "@/components/analytics/ClientSideTracker";
 import { GlobalSidebar } from "@/components/sidebar";
+import Dreamshaper from "@/components/welcome/featured/Dreamshaper";
+import { usePrivy } from "@privy-io/react-auth";
 import { SidebarProvider } from "@repo/design-system/components/ui/sidebar";
 import { cn } from "@repo/design-system/lib/utils";
 import { useEffect, useState } from "react";
-import { usePrivy } from "@privy-io/react-auth";
+import { useOnboard } from "../OnboardContext";
 
 export default function MainExperience() {
   const { cameraPermission, currentStep } = useOnboard();
@@ -34,7 +35,8 @@ export default function MainExperience() {
     >
       <GlobalSidebar>
         <div className="flex h-screen md:h-auto md:min-h-[calc(100vh-2rem)] flex-col px-2 md:px-6">
-          <DayDreamContent />
+          <Dreamshaper />
+          <ClientSideTracker eventName="home_page_view" />
         </div>
       </GlobalSidebar>
     </SidebarProvider>
