+++ apps/app/components/welcome/featured/ShareModal.tsx
@@ -1,6 +1,8 @@
 "use client";
 
-import React, { useState, useRef, useEffect } from "react";
+import { usePromptVersionStore } from "@/hooks/usePromptVersionStore";
+import track from "@/lib/track";
+import { usePrivy } from "@privy-io/react-auth";
 import { Button } from "@repo/design-system/components/ui/button";
 import {
   Dialog,
@@ -11,24 +13,27 @@ import {
 } from "@repo/design-system/components/ui/dialog";
 import { Input } from "@repo/design-system/components/ui/input";
 import { Copy, Loader2 } from "lucide-react";
+import { useEffect, useRef, useState } from "react";
 import { toast } from "sonner";
-import track from "@/lib/track";
-import { usePromptVersionStore } from "@/hooks/usePromptVersionStore";
-import { usePrivy } from "@privy-io/react-auth";
+import {
+  useDreamshaperStore,
+  useShareLink,
+} from "../../../hooks/useDreamshaper";
 
-interface ShareModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  createShareLink: () => Promise<{ error: string | null; url: string | null }>;
-  streamId?: string | null;
+export function useShareModal() {
+  const [open, setOpen] = useState(false);
+
+  return {
+    open,
+    setOpen,
+    openModal: () => setOpen(true),
+    closeModal: () => setOpen(false),
+  };
 }
 
-export function ShareModal({
-  isOpen,
-  onClose,
-  createShareLink,
-  streamId,
-}: ShareModalProps) {
+export const ShareModalContent = () => {
+  const { createShareLink } = useShareLink();
+  const { stream } = useDreamshaperStore();
   const { promptVersion } = usePromptVersionStore();
   const { authenticated } = usePrivy();
   const [isCreating, setIsCreating] = useState(false);
@@ -50,7 +55,7 @@ export function ShareModal({
         setShareUrl(result.url);
         track("daydream_share_link_created", {
           is_authenticated: authenticated,
-          stream_id: streamId,
+          stream_id: stream?.id,
         });
       }
     } catch (error) {
@@ -67,59 +72,57 @@ export function ShareModal({
       toast.success("Link copied to clipboard!");
       track("daydream_share_link_copied", {
         is_authenticated: authenticated,
-        stream_id: streamId,
+        stream_id: stream?.id,
       });
     }
   };
 
   return (
-    <Dialog open={isOpen} onOpenChange={onClose}>
-      <DialogContent className="mx-auto left-[50%] -translate-x-[50%] w-[calc(100%-2rem)] bg-background border border-border rounded-xl p-5 max-w-sm shadow-lg">
-        <DialogHeader className="pb-2">
-          <DialogTitle className="text-xl">Share your experience</DialogTitle>
-          <DialogDescription className="pt-2 text-sm text-muted-foreground">
-            Your personalization, prompt, and parameters will be shared.
-            <br />
-            <br />
-            Anyone with this link will see the exact same experience you've
-            created.
-          </DialogDescription>
-        </DialogHeader>
-        <div className="flex flex-col gap-4 mb-2 mt-3">
-          {shareUrl ? (
-            <div className="flex items-center gap-2">
-              <Input
-                ref={inputRef}
-                value={shareUrl}
-                readOnly
-                className="flex-1 rounded-md border-muted-foreground/20 focus-visible:ring-offset-1 text-sm"
-              />
-              <Button
-                size="icon"
-                onClick={handleCopyLink}
-                className="rounded-md hover:bg-primary/90"
-              >
-                <Copy className="h-4 w-4" />
-              </Button>
-            </div>
-          ) : (
+    <DialogContent className="mx-auto left-[50%] -translate-x-[50%] w-[calc(100%-2rem)] bg-background border border-border rounded-xl p-5 max-w-sm shadow-lg">
+      <DialogHeader className="pb-2">
+        <DialogTitle className="text-xl">Share your experience</DialogTitle>
+        <DialogDescription className="pt-2 text-sm text-muted-foreground">
+          Your personalization, prompt, and parameters will be shared.
+          <br />
+          <br />
+          Anyone with this link will see the exact same experience you've
+          created.
+        </DialogDescription>
+      </DialogHeader>
+      <div className="flex flex-col gap-4 mb-2 mt-3">
+        {shareUrl ? (
+          <div className="flex items-center gap-2">
+            <Input
+              ref={inputRef}
+              value={shareUrl}
+              readOnly
+              className="flex-1 rounded-md border-muted-foreground/20 focus-visible:ring-offset-1 text-sm"
+            />
             <Button
-              onClick={handleCreateLink}
-              disabled={isCreating}
-              className="w-full rounded-md hover:bg-primary/90"
+              size="icon"
+              onClick={handleCopyLink}
+              className="rounded-md hover:bg-primary/90"
             >
-              {isCreating ? (
-                <>
-                  <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-                  Creating link...
-                </>
-              ) : (
-                "Create link"
-              )}
+              <Copy className="h-4 w-4" />
             </Button>
-          )}
-        </div>
-      </DialogContent>
-    </Dialog>
+          </div>
+        ) : (
+          <Button
+            onClick={handleCreateLink}
+            disabled={isCreating}
+            className="w-full rounded-md hover:bg-primary/90"
+          >
+            {isCreating ? (
+              <>
+                <Loader2 className="mr-2 h-4 w-4 animate-spin" />
+                Creating link...
+              </>
+            ) : (
+              "Create link"
+            )}
+          </Button>
+        )}
+      </div>
+    </DialogContent>
   );
-}
+};
