+++ apps/app/hooks/useStreamStatus.ts
@@ -1,3 +1,4 @@
+import useSWR from "swr";
 import { usePrivy } from "@privy-io/react-auth";
 import { useSearchParams } from "next/navigation";
 import { useEffect, useState, useRef } from "react";
@@ -13,20 +14,64 @@ export enum StreamStatus {
 }
 
 export const useStreamStatus = (
-  streamId: string,
+  streamId?: string,
   requireUser: boolean = true,
 ) => {
   const { ready, user } = usePrivy();
-  const [status, setStatus] = useState<StreamStatus | null>(null);
-  const [fullResponse, setFullResponse] = useState<any>(null);
-  const [loading, setLoading] = useState<boolean>(true);
-  const [error, setError] = useState<string | null>(null);
   const [maxStatusLevel, setMaxStatusLevel] = useState(0);
-  const failureCountRef = useRef(0);
   const orchestratorFailureCountRef = useRef(0);
-  const intervalIdRef = useRef<NodeJS.Timeout | null>(null);
   const searchParams = useSearchParams();
 
+  const shouldFetch = ready && (!requireUser || !!user) && !!streamId;
+
+  const getUrl = () => {
+    if (!shouldFetch) return null;
+    return searchParams.get("gateway")
+      ? `/api/streams/${streamId}/status?gateway=${searchParams.get("gateway")}`
+      : `/api/streams/${streamId}/status`;
+  };
+
+  const fetcher = async (url: string) => {
+    const res = await fetch(url, {
+      headers: {
+        "cache-control": "no-cache",
+        pragma: "no-cache",
+      },
+    });
+
+    if (!res.ok) {
+      throw new Error(
+        `Failed to fetch stream status: ${res.status} ${res.statusText}`,
+      );
+    }
+
+    const { success, error, data } = await res.json();
+    if (!success || !data) {
+      throw new Error(error ?? "No stream data returned from API");
+    }
+
+    return data;
+  };
+
+  const { data: fullResponse, error } = useSWR(getUrl, fetcher, {
+    refreshInterval: BASE_POLLING_INTERVAL,
+    dedupingInterval: 1000,
+    revalidateOnFocus: true,
+    onSuccess: (data: any) => {
+      if (
+        data?.gateway_status?.last_error?.startsWith(
+          "no orchestrators available within",
+        )
+      ) {
+        orchestratorFailureCountRef.current += 1;
+      } else {
+        orchestratorFailureCountRef.current = 0;
+      }
+    },
+  });
+
+  const status = fullResponse?.state || StreamStatus.Unknown;
+  const loading = !error && !fullResponse;
   const capacityReached = orchestratorFailureCountRef.current >= 5;
 
   const isLive =
@@ -71,86 +116,11 @@ export const useStreamStatus = (
     }
   };
 
-  const triggerError = (errMsg: string) => {
-    setError(errMsg);
-    setLoading(false);
-  };
-
-  const resetPollingInterval = () => {
-    // Optinal func to adjust the polling interval.
-  };
-
-  useEffect(() => {
-    if (!ready || (requireUser && !user)) {
-      console.log(
-        "[useStreamStatus] Not ready or user required but not detected, aborting status fetch.",
-      );
-      return;
-    }
-    if (!streamId) return;
-
-    const fetchStatus = async () => {
-      try {
-        const url = searchParams.get("gateway")
-          ? `/api/streams/${streamId}/status?gateway=${searchParams.get("gateway")}`
-          : `/api/streams/${streamId}/status`;
-
-        const res = await fetch(url, {
-          headers: {
-            "cache-control": "no-cache",
-            pragma: "no-cache",
-          },
-        });
-        if (!res.ok) {
-          triggerError(
-            `Failed to fetch stream status: ${res.status} ${res.statusText}`,
-          );
-          return;
-        }
-
-        const { success, error, data } = await res.json();
-        if (!success || !data) {
-          triggerError(error ?? "No stream data returned from API");
-          return;
-        }
-        setFullResponse(data);
-        setStatus((data?.state as StreamStatus) || StreamStatus.Unknown);
-        setError(null);
-
-        if (
-          data?.gateway_status?.last_error?.startsWith(
-            "no orchestrators available within",
-          )
-        ) {
-          orchestratorFailureCountRef.current += 1;
-        } else {
-          orchestratorFailureCountRef.current = 0;
-        }
-
-        failureCountRef.current = 0;
-        resetPollingInterval();
-      } catch (err: any) {
-        triggerError(err.message);
-        setLoading(false);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchStatus();
-    const intervalId = setInterval(fetchStatus, BASE_POLLING_INTERVAL);
-    intervalIdRef.current = intervalId;
-
-    return () => {
-      if (intervalIdRef.current) clearInterval(intervalIdRef.current);
-    };
-  }, [ready, streamId, requireUser, user]);
-
   return {
     status,
     fullResponse,
     loading,
-    error,
+    error: error?.message || null,
     isLive,
     statusMessage: getStatusMessage(),
     capacityReached,
