+++ apps/app/package.json
@@ -42,6 +42,7 @@
     "react-dom": "18.3.1",
     "react-textarea-autosize": "^8.5.7",
     "sonner": "^1.7.0",
+    "swr": "^2.3.3",
     "zod": "^3.24.1",
     "zustand": "^5.0.3"
   },
